[FIX] survey: comment as answer not working on roaming multiple choice
Issue:
When answering a question with a comment in multiple choice with roaming
activated for the survey, the UI will display a warning message that says
the question requires an answer.

Cause:
The backend creates a skipped record if none of the pre-created answers
is chosen.

Solution:
Don't create a skipped record if a comment counts as an answer and a
comment is provided.

Added validation of input and unittests

Task-5062984

X-original-commit: 11b2f97a4b33b203868384edfd8042bc69ef1ba1
Part-of: odoo/odoo#230658
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Jérôme Belpaire (beje) <beje@odoo.com>