[REF] web: Hoot - rename 'exact' to 'count'
This commit renames the 'exact' option given to queryX helpers to
'count', as to better reflect what it specifies, i.e. returning nodes
if their count matches the given one.

This commit also adds the ":count()" pseudo-class to be used in
selectors, for consistency with other options which are also pseudo-
classes.

closes odoo/odoo#230556

Related: odoo/enterprise#96647
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>